pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dreams_core"
version = "0.0.1"
authors = [
  { name="tentabs", email="tentabs00@gmail.com" },
]
description = "brought to you by the dreamslabs discord community"
readme = "README.md"
requires-python = ">=3.11"
exclude = [
  "**/*.pyc",
  ".gitignore",
  "requirements.txt",
  ".DS_Store"
]
dependencies = [
  "requests>=2.25.1",
  "pandas>=1.2.0",
  "numpy>=1.19.5",
  "google-auth>=1.24.0",
  "google-cloud-secret-manager>=2.3.0",
  "google-cloud-bigquery>=2.13.1",
  "google-cloud-storage>=1.37.1"
]

[project.urls]
Community = "https://discord.gg/dreamcrypto"
Github = "https://github.com/dreams-labs/core-functions"

# Classifiers to help categorize the project
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Dreams Crypto Community Devs",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]